Move around definitions in sc_network
